Update forecast.py
Change hourly forecast precip to float from int. When getting data in inches, int would round up/down to 0.04 with the unit conversion.

+++ weatherflow4py/models/rest/forecast.py
@@ -275,7 +275,7 @@ class ForecastHourly:
 
     local_day: int
     local_hour: int
-    precip: int
+    precip: float
     precip_probability: int
     # precip_type: PrecipType
     precip_type: PrecipType = field(
